requirements.txt
einops==0.3.2
hydra-core==1.2.0
numpy
opencv-python-headless
Pillow

pytorch-lightning==1.6.3
tensorflow==2.7.0
timm==0.5.4
torch==1.9.0
torchaudio==0.9.0
torchmetrics==0.8.2
torchvision==0.10.0
tqdm==4.64.0
setuptools==58.2.0